refactor[w2.1]:move uploading image to patient registration page
